consolidating tests to ignore_test file
